requirements.txt
# python == 3.10.11
absl-py==2.1.0
acme==2.9.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.59
botocore==1.34.59
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
distro==1.9.0
dlib==19.24.2
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
gdown==5.1.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.6
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
josepy==1.14.0
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
mtcnn==0.1.1
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parsedatetime==2.6
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rsa==4.9
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1